pydrake.util: Rename `module_shim` to `deprecation`
