+++ bindings/pydrake/util/BUILD.bazel
@@ -52,8 +52,8 @@ drake_py_library(
 )
 
 drake_py_library(
-    name = "module_shim_py",
-    srcs = ["module_shim.py"],
+    name = "deprecation_py",
+    srcs = ["deprecation.py"],
     imports = PACKAGE_INFO.py_imports,
 )
 
@@ -159,7 +159,7 @@ PY_LIBRARIES = [
     ":cpp_param_py",
     ":cpp_template_py",
     ":module_py",
-    ":module_shim_py",
+    ":deprecation_py",
     ":pybind11_version_py",
 ]
 
@@ -200,22 +200,22 @@ drake_cc_googletest(
 )
 
 drake_py_library(
-    name = "module_shim_example",
+    name = "deprecation_example",
     testonly = 1,
     srcs = [
-        "test/module_shim_example/__init__.py",
-        "test/module_shim_example/sub_module.py",
+        "test/deprecation_example/__init__.py",
+        "test/deprecation_example/sub_module.py",
     ],
     imports = ["test"],
-    deps = [":module_shim_py"],
+    deps = [":deprecation_py"],
 )
 
 drake_py_test(
-    name = "module_shim_test",
-    srcs = ["test/module_shim_test.py"],
+    name = "deprecation_test",
+    srcs = ["test/deprecation_test.py"],
     allow_import_unittest = True,
     deps = [
-        ":module_shim_example",
+        ":deprecation_example",
     ],
 )
