+++ bindings/pydrake/util/all.py
@@ -2,4 +2,4 @@
 # - `cpp_param` does not offer public Drake symbols.
 # - `cpp_template` does not offer public Drake symbols.
 from .eigen_geometry import *
-# - `module_shim` does not offer public Drake symbols.
+# - `deprecation` does not offer public Drake symbols.
