+++ bindings/pydrake/util/deprecation.py
