+++ bindings/pydrake/util/test/deprecation_example/__init__.py
@@ -1,4 +1,4 @@
-from pydrake.util.module_shim import ModuleShim
+from pydrake.util.deprecation import ModuleShim
 
 value = 1
